uart: rewrite the whole PL011 driver.
The strange dead loop bug make me realize that I don't
have a thorough understanding of how PL011 works (or I
forget it...)

So I read the PL011 TRM r1p5 and the Xinu OS PL011 impl.
And rewrite the whole PL011 driver. The ATF-M PL011 code
is re-used, which is basically a bunch of register access
wrapper functions.

Now I have a better understanding of PL011. And I wrote
a lot of comment about that. It's a fascinating experience.

The new code is meant to work in FIFO mode or single byte mode.
But QEMU PL011 model ignores the FIFO. So I would have to
get a real board to test.

But so far it works fine. And the dead loop bug should have
been fixed now.

But I found another issue with the LCD display when handling
the \r\n sequence. The char at the line head will be erased.

I will fix it later.

Signed-off-by: smwikipedia <smwikipedia@163.com>