cm_os: redesign the folder structure 2
Most important things are below:

"drivers" provides an API for a peripheral instance.
The API takes the register block as a parameter.

"boards" provide an API to the OS for all the available
peripheral instances on a board. The API takes a logical
index of the peripheral instance as a parameter.

And give up adopting CMSIS code as is. Such as Driver_USART.h.
Only use it as a reference.

And add a .clang-format file.

Signed-off-by: smwikipedia <smwikipedia@163.com>